fix PR, improved title nav component
