+++ docs-italia-theme/data/_config.yml
@@ -1,6 +1,6 @@
 # Custom vars
 theme_color: "#0066cc"
-css_theme: t-Pac doc-Italia
+css_theme: doc-Italia
 owner_short: AgID
 owner_full: AgID
 owner_link: http://www.agid.gov.it
