+++ docs-italia-theme/layouts/footer.html
@@ -1,5 +1,5 @@
 <div class="bg-primary--darken">
-    <footer class="Footer container-fluid">
+    <footer class="container-fluid">
 
       <div class="row pt-5 pb-5 pl-4 pl-md-5 pr-4 pr-md-5">
