+++ docs-italia-theme/layouts/header.html
@@ -1,4 +1,4 @@
-<header class="Header">
+<header>
 
     <!-- <a class="js-fr-offcanvas-open u-block" href="#network" aria-controls="network"
       aria-label="{{ t.toggle_network_menu_label }}" title="{{ t.toggle_network_menu_label }}">
