--- docs-italia-theme/layouts/page_project.html
@@ -1,71 +1,0 @@
-{% include 'layouts/head.html' %}
-
-<div class="u-background-grey-80">
-  <div id="main">
-    {% include 'layouts/header.html' %}
-    {% include 'layouts/network.html' %}
-
-    {% from 'layouts/card.html' import card with context %}
-
-    <div class="u-background-gradient-t-b-blue-50 u-textLeft">
-      <div class="Grid Grid--alignMiddle u-layout-wide u-layoutCenter u-padding-r-all u-posRelative u-zindex-30">
-        <div class="Grid-cell u-sm-size2of4">
-          <h1 class="u-color-white u-minHeight-80">
-            ANPR <span class="u-textWeight-200">- Anagrafe nazionale popolazione residente</span>
-          </h1>
-          <span class="u-text-r-xs u-padding-r-top u-color-10 u-layout-prose">Documenti Disponibili</span>
-        </div>
-      </div>
-      <hr class="Separator-room">
-      <hr class="Separator Separator--up u-background-grey-15">
-    </div>
-
-    <div class="u-background-grey-15 u-padding-r-bottom u-textLeft">
-      <section class="u-padding-r-all u-layout-wide Grid u-layoutCenter u-posRelative u-zindex-30">
-
-        <div class="u-layout-wide u-pullUp">
-          <div class="Grid u-flex u-flexRowReverse">{# introductory text on the right side for larger screens #}
-            <div class="Grid-cell u-sm-size2of4 u-md-size2of4 u-lg-size2of4 u-pullUp">
-              <div class="Grid-cell  u-sm-before1of4 u-sm-size3of4 u-md-before1of4 u-md-size3of4 u-lg-before1of4 u-lg-size3of4 u-margin-r-bottom u-layout-matchHeight" >
-                <div class="u-nbfc u-flexWrap u-flex u-color-grey-30 u-xs-padding-all-none u-borderShadow-m u-xs-borderShadow-none u-borderRadius-m u-background-white u-sizeFill">
-                  <div class="u-flexWrap u-flex u-flexAlignSelfStretch u-sizeFill u-padding-r-all">
-                    <div class="u-sizeFull u-padding-r-all u-xs-padding-all-none">
-                      <h4 class="u-text-h4 u-color-grey-50 u-padding-bottom-xs">About the project</h4>
-                      <p class="Content u-text-p u-textSecondary u-padding-r-bottom">
-                        Duis lobortis dui non justo vulputate luctus. Donec lorem nunc, tempus sodales urna in, porttitor condimentum nisl. Vivamus non est egestas, tristique est id, rhoncus orci.
-                      </p>
-                      <h4 class="u-text-h4 u-color-grey-50 u-padding-bottom-xs">More on</h4>
-                      <ul>
-                        <li>
-                          <a href="#" class="u-color-50 u-text-r-xxs u-textWeight-600 u-textClean">per saperne di più</a>
-                        </li><li>
-                          <a href="#" class="u-color-50 u-text-r-xxs u-textWeight-600 u-textClean">per saperne di più</a>
-                        </li>
-                      </ul>
-                    </div>
-                  </div>
-                </div>
-              </div>
-            </div>
-            <div class="Grid-cell u-sm-size2of4 u-md-size2of4 u-lg-size2of4">
-                <div class="Grid Grid--withGutter">
-                  {% for subproject in subproject_data %}
-                    <div class="Grid-cell u-md-size1of2 u-lg-size1of2 u-margin-r-bottom u-layout-matchHeight" id="subproject-1">
-                        {{ card(subproject.name, subproject.title, subproject.url, subproject.description, subproject.version) }}
-                    </div>
-                  {% endfor %}
-                </div>
-              </div>
-          </div>
-        </div>
-      </section>
-    </div>
-  </div>
-</div>
-
-{% include 'layouts/scrolltop.html' %}
-{% include 'layouts/footer.html' %}
-{% include 'layouts/scripts.html' %}
-
-</body>
-</html>
