+++ docs-italia-theme/layouts/scripts.html
@@ -18,15 +18,4 @@
 {# RTD hosts this file, so just load on non RTD builds #}
 {% if not READTHEDOCS %}
 <script src="{{ pathto('_static/js/theme.js', 1) }}"></script>
-<script src="{{ pathto('_static/js/custom.js', 1) }}"></script>
 {% endif %}
-
-
-<!-- {# STICKY NAVIGATION #}
-{% if theme_sticky_navigation %}
-<script type="text/javascript">
-    jQuery(function() {
-        SphinxRtdTheme.StickyNav.enable();
-    });
-</script>
-{% endif %} -->
