+++ docs-italia-theme/static/js/theme.js
@@ -1,175 +1,256 @@
+
 require=(function(){function e(t,n,r){function s(o,u){if(!n[o]){if(!t[o]){var a=typeof require=="function"&&require;if(!u&&a)return a(o,!0);if(i)return i(o,!0);var f=new Error("Cannot find module '"+o+"'");throw f.code="MODULE_NOT_FOUND",f}var l=n[o]={exports:{}};t[o][0].call(l.exports,function(e){var n=t[o][1][e];return s(n?n:e)},l,l.exports,e,t,n,r)}return n[o].exports}var i=typeof require=="function"&&require;for(var o=0;o<r.length;o++)s(r[o]);return s}return e})()({"docs-italia-theme":[function(require,module,exports){
-var jQuery = (typeof(window) != 'undefined') ? window.jQuery : require('jquery');
+// Keywords Tolltip
+var ThemeToolTip = (function ($) {
+  var that;
 
-// Sphinx theme nav state
-function ThemeNav() {
+  return {
 
-    var nav = {
-        navBar: null,
-        win: null,
-        winScroll: false,
-        winResize: false,
-        linkScroll: false,
-        winPosition: 0,
-        winHeight: null,
-        docHeight: null,
-        isRunning: false
-    };
+    $: {
+      btn: $('.pull-quote .reference.internal'),
+      docWindow: $( window ),
+      $body: $('body'),
+      toolTipArray: []
+    },
 
-    nav.enable = function () {
-        var self = this;
+    init: function() {
+      that = this.$;
+      ThemeToolTip.addAttribute();
+    },
 
-        if (!self.isRunning) {
-            self.isRunning = true;
-            jQuery(function ($) {
-                self.init($);
+    // Add atribute to keywords btn for enable tooltip
+    addAttribute: function() {
+      that.btn.each(function(index) {
+        var title = $(this).find('span').html();
+        $(this).attr('data-toggle','tooltip').attr('data-html','true').attr('title',title).attr('data-ref',index);
+        that.toolTipArray.push(new ThemeToolTip.setData($(this),index));
+      });
+      ThemeToolTip.addhandler();
+    },
 
-                self.reset();
-                self.win.on('hashchange', self.reset);
+    // Set array whith keywords btn info
+    setData: function(item , index) {
+      this.btn = item;
+      this.term = item.attr('href');
+      this.title = $(this.term).html();
+      this.body = $(this.term).next('dd').html();
+      this.ref = index
+    },
 
-                // Set scroll monitor
-                self.win.on('scroll', function () {
-                    if (!self.linkScroll) {
-                        self.winScroll = true;
-                    }
-                });
-                setInterval(function () {
-                    self.onScroll();
-                }, 25);
+    // Enable toottip custom
+    addhandler: function() {
+      for (var index = 0; index < that.toolTipArray.length; ++index) {
+        var toolTipTemplate = "<div class='tooltip tooltip--active doc-tooltip' role='tooltip'><div class='tooltip__wrap'>" +
+            "<button type='button' class='tooltip__close-btn' data-ref=" + that.toolTipArray[index].ref + "></button>" +
+            "<h2 class='tooltip__title'>" + that.toolTipArray[index].title + "</h2>" +
+            "<p class='tooltip__content'>" + that.toolTipArray[index].body + "</p>" +
+            "</div></div>";
+            btn = that.toolTipArray[index].btn;
+        btn.tooltip({template:toolTipTemplate,trigger:'click', placement:'top',offset:'115px , 40px'});
+      };
 
+      // Close toolTip from button.
+      $(document).on('click','.tooltip__close-btn', function(){
+        var ref = $(this).attr('data-ref');
+            btn = that.btn.filter(function(){
+              return ($(this).attr("data-ref") == ref)
+             });
+        btn.tooltip('hide');
+      });
 
-                // Set resize monitor
-                self.win.on('resize', function () {
-                    self.winResize = true;
-                });
-                setInterval(function () {
-                    if (self.winResize) self.onResize();
-                }, 25);
-                self.onResize();
-            });
+      // Close other btn when one is active.
+      that.btn.on('click' , function(){
+        that.btn.not($(this)).tooltip('hide');
+      });
+
+      // Hide tool tip on resize
+      that.docWindow.resize(function() {
+        that.btn.tooltip('hide');
+      });
+
+      // Hide tool tip when body is clickked
+      that.$body.on('click' , function(event) {
+        if( !$('.tooltip').has(event.target).length > 0 && !$(event.target).is($('.std-term'))) {
+          that.btn.tooltip('hide');
         }
-        ;
-    };
-
-    nav.init = function ($) {
-        var doc = $(document),
-            self = this;
-
-        this.navBar = $('div.wy-side-scroll:first');
-        this.win = $(window);
-
-        // Set up javascript UX bits
-        $(document)
-        // Shift nav in mobile when clicking the menu.
-            .on('click', "[data-toggle='wy-nav-top']", function () {
-                $("[data-toggle='wy-nav-shift']").toggleClass("shift");
-                $("[data-toggle='rst-versions']").toggleClass("shift");
-            })
-
-            // Nav menu link click operations
-            .on('click', ".wy-menu-vertical .current ul li a", function () {
-                var target = $(this);
-                // Close menu when you click a link.
-                $("[data-toggle='wy-nav-shift']").removeClass("shift");
-                $("[data-toggle='rst-versions']").toggleClass("shift");
-                // Handle dynamic display of l3 and l4 nav lists
-                self.toggleCurrent(target);
-                self.hashChange();
-            })
-            .on('click', "[data-toggle='rst-current-version']", function () {
-                $("[data-toggle='rst-versions']").toggleClass("shift-up");
-            })
-
-        // Make tables responsive
-        $("table.docutils:not(.field-list)")
-            .wrap("<div class='wy-table-responsive'></div>");
-
-        // Add expand links to all parents of nested ul
-        $('.wy-menu-vertical ul').not('.simple').siblings('a').each(function () {
-            var link = $(this);
-            expand = $('<span class="toctree-expand"></span>');
-            expand.on('click', function (ev) {
-                self.toggleCurrent(link);
-                ev.stopPropagation();
-                return false;
-            });
-            link.prepend(expand);
-        });
-    };
-
-    nav.reset = function () {
-        // Get anchor from URL and open up nested nav
-        var anchor = encodeURI(window.location.hash);
-        if (anchor) {
-            try {
-                var link = $('.wy-menu-vertical')
-                    .find('[href="' + anchor + '"]');
-                // If we didn't find a link, it may be because we clicked on
-                // something that is not in the sidebar (eg: when using
-                // sphinxcontrib.httpdomain it generates headerlinks but those
-                // aren't picked up and placed in the toctree). So let's find
-                // the closest header in the document and try with that one.
-                if (link.length === 0) {
-                    var doc_link = $('.document a[href="' + anchor + '"]');
-                    var closest_section = doc_link.closest('div.section');
-                    // Try again with the closest section entry.
-                    link = $('.wy-menu-vertical')
-                        .find('[href="#' + closest_section.attr("id") + '"]');
-
-                }
-                $('.wy-menu-vertical li.toctree-l1 li.current')
-                    .removeClass('current');
-                link.closest('li.toctree-l2').addClass('current');
-                link.closest('li.toctree-l3').addClass('current');
-                link.closest('li.toctree-l4').addClass('current');
-            }
-            catch (err) {
-                console.log("Error expanding nav for anchor", err);
-            }
-        }
-    };
-
-    nav.onScroll = function () {
-        this.winScroll = false;
-        var newWinPosition = this.win.scrollTop(),
-            winBottom = newWinPosition + this.winHeight,
-            navPosition = this.navBar.scrollTop(),
-            newNavPosition = navPosition + (newWinPosition - this.winPosition);
-        if (newWinPosition < 0 || winBottom > this.docHeight) {
-            return;
-        }
-        this.navBar.scrollTop(newNavPosition);
-        this.winPosition = newWinPosition;
-    };
-
-    nav.onResize = function () {
-        this.winResize = false;
-        this.winHeight = this.win.height();
-        this.docHeight = $(document).height();
-    };
-
-    nav.hashChange = function () {
-        this.linkScroll = true;
-        this.win.one('hashchange', function () {
-            this.linkScroll = false;
-        });
-    };
-
-    nav.toggleCurrent = function (elem) {
-        var parent_li = elem.closest('li');
-        parent_li.siblings('li.current').removeClass('current');
-        parent_li.siblings().find('li.current').removeClass('current');
-        parent_li.find('> ul li.current').removeClass('current');
-        parent_li.toggleClass('current');
+      });
     }
+  }
+})(jQuery);
 
-    return nav;
-};
+ThemeToolTip.init();
 
-module.exports.ThemeNav = ThemeNav();
 
-if (typeof(window) != 'undefined') {
-    window.SphinxRtdTheme = {StickyNav: module.exports.ThemeNav};
-}
+/// Modify DOM via JS.
+var ThemeMarkupModifier = (function ($) {
+  var that;
 
-},{"jquery":"jquery"}]},{},["docs-italia-theme"]);
+  return {
+
+    $: {
+      title: $('#doc-content h1, #doc-content h2, #doc-content h3'),
+      $table: $('table.docutils')
+    },
+
+    init: function() {
+      that = this.$;
+      ThemeMarkupModifier.titleModifier();
+      ThemeMarkupModifier.tableModifier()
+    },
+
+    titleModifier: function() {
+      that.title.each(function(index) {
+
+        var $element = $(this),
+            title = $element.html(),
+            number = ThemeMarkupModifier.startingNumber(title);
+
+        $element.wrap('<div class="chapter-header clearfix"><div class="title-wrap">');
+        if( !ThemeMarkupModifier.dotNumberValidator(number) ) {
+          $element.html(title.replace(number , '<span class="title__chapter">' + number + '</span>'));
+          $element.addClass('has-nav');
+          $element.closest('.title-wrap').append('<span class="title__background">');
+        }
+      })
+    },
+
+    startingNumber: function(string) {
+      var startingNumber=string.substr(0 , string.indexOf(' '))
+      return startingNumber;
+    },
+
+    dotNumberValidator: function(string) {
+      var regex = /^[0-9.]+$/;
+      return !regex.test(string)
+    },
+
+    tableModifier: function() {
+      that.$table.wrap('<div class="table-responsive">');
+    }
+  }
+})(jQuery);
+
+ThemeMarkupModifier.init();
+
+
+/// Paragraph navigation
+var ThemeChapterNav = (function ($) {
+  var that;
+
+  return {
+
+    $: {
+      $title: $('h1.has-nav, h2.has-nav, h3.has-nav'),
+      $body: $('body'),
+      $html: $('html'),
+      $window: $(window)
+    },
+
+    init: function() {
+      that = this.$;
+      that.$title.each(function(index) {
+        $element = $(this);
+        ThemeChapterNav.addNav($element);
+      });
+      ThemeChapterNav.addHandler();
+    },
+
+    addNav: function(element) {
+      var nav = "<div class='chapter-nav'><div class='chapter-nav__wrap'>" +
+          "<ul class='chapter-nav__list chapter-nav__list--visible'>" +
+          "<li class='chapter-nav__item'>" +
+          "<span class='u-text-r-l Icon Icon-comment'></span>" +
+          "<button type='button' class='chapter-link'><span>4</span> commenti</button></li>" +
+          "<li class='chapter-nav__item'>" +
+          "<span class='u-text-r-l Icon Icon-more-items'></span>" +
+          "<button type='button' class='chapter-link chapter-link--expand'>Vedi Azioni</button></li>" +
+          "<li class='chapter-nav__item'>" +
+          "<span class='u-text-r-l Icon Icon-more-items'></span>" +
+          "<button type='button' class='chapter-link chapter-link--expand'>Altre Azioni</button></li></ul>" +
+          "<div class='chapter-nav__list--hidden'>" +
+          "<ul class='chapter-nav__list'>" +
+          "<li class='chapter-nav__item'>" +
+          "<span class='u-text-r-l Icon Icon-link'></span>" +
+          "<button type='button' class='chapter-link'>Copia link</button></li>" +
+          "<li class='chapter-nav__item'>" +
+          "<span class='u-text-r-l Icon Icon-share'></span>" +
+          "<button type='button' class='chapter-link'>Confronta versioni</button></li>" +
+          "<li class='chapter-link chapter-nav__item'>" +
+          "<span class='u-text-r-l Icon Icon-share'></span>" +
+          "<button type='button' class='chapter-link'>Condividi</button>" +
+          "</li></ul></div></div></div>";
+      container = element.closest('.chapter-header');
+      container.append(nav);
+    },
+
+    addHandler: function(element) {
+      var $expanded = $('.chapter-link--expand');
+
+      if (!that.$html.hasClass('touch')) {
+        $('.chapter-header').on('mouseover' , function(){
+          if( that.$window.outerWidth() > 576 ) {
+            var $nav = $(this).closest('.chapter-header').find('.chapter-nav__list--hidden'),
+            $wrap = $(this).find('.title-wrap');
+            lineHeight = parseInt($(this).find('.has-nav').css('line-height'))+4;
+
+            $nav.addClass('active');
+            $wrap.addClass('active');
+            $wrap.find('.title__background').css('height',lineHeight);
+          }
+        });
+        $('.chapter-header').on('mouseout' , function(){
+          if( that.$window.outerWidth() > 576 ) {
+            var $nav = $(this).closest('.chapter-header').find('.chapter-nav__list--hidden'),
+                $wrap = $(this).find('.title-wrap');
+            $nav.removeClass('active');
+            $wrap.removeClass('active');
+          }
+        });
+      }
+
+      else if (that.$html.hasClass('touch')) {
+        // Close lightbox on click ( toch monitor)
+        $('.chapter-nav__list--hidden').on('click' , function(event){
+          if( $(event.target).is( $('.chapter-nav__list--hidden'))){
+            var $wrap = $(this).closest('.chapter-header').find('.title-wrap');
+            $(this).removeClass('active');
+            that.$body.removeClass('no-scroll');
+            $wrap.removeClass('active');
+          }
+        });
+
+        // Open nav ( toch monitor)
+        $expanded.on('click' , function(){
+          var $nav = $(this).closest('.chapter-header').find('.chapter-nav__list--hidden'),
+              $wrap = $(this).closest('.chapter-header').find('.title-wrap'),
+              lineHeight = parseInt($(this).closest('.chapter-header').find('.has-nav').css('line-height'))+4;
+          if( $nav.hasClass('active') ) {
+            $nav.removeClass('active');
+            $wrap.removeClass('active');
+          } else {
+            $nav.addClass('active');
+            $wrap.addClass('active');
+            $wrap.find('.title__background').css('height',lineHeight);
+            if(that.$window.outerWidth() <= 576) {
+              that.$body.addClass('no-scroll');
+            }
+          }
+        });
+      }
+
+      // Close when bosy is clikked.
+      that.$body.on('click' , function(){
+        var $nav = that.$title.closest('.chapter-header').find('.chapter-nav__list--hidden');
+        if( !$('.chapter-header').has(event.target).length > 0 ) {
+          $('.chapter-nav__list--hidden').removeClass('active');
+          $('.title-wrap').removeClass('active');
+        }
+      });
+
+    }
+  }
+})(jQuery);
+
+ThemeChapterNav.init();
+
+},{}]},{},["docs-italia-theme"]);
