+++ sass/abstract/_variables.sass
@@ -16,3 +16,5 @@ $red: #d83750
 // Font
 $stack-font: 'Lora', serif
 $titillium-font: 'Titillium Web', 'Helvetica Neue', Helvetica, Arial, sans-serif
+
+$chapter-nav-width: 200px
