+++ sass/layout/_content-nav.sass
@@ -1,18 +1,38 @@
-.chapter-nav
-  position: relative
-
 .chapter-nav__list--hidden
-  visibility: hidden
-  opacity: 0
-  transition: visibility .35s,  opacity .35s
-  &.active
-    visibility: visible
-    opacity: 1
+  @media (max-width: 575px)
+    display: none
+    position: fixed
+    top: 0
+    bottom: 0
+    left: 0
+    right: 0
+    background-color: rgba($black , .6)
+    .chapter-nav__list
+      background-color: $white
+      position: absolute
+      top: 50%
+      left: 50%
+      transform: translate(-50% , -50%)
+      padding: 2rem
+      border-radius: .5rem
+      width: 100%
+      max-width: 250px
+    &.active
+      display: block
+
+  @media (min-width: 576px)
+    visibility: hidden
+    opacity: 0
+    transition: visibility .35s,  opacity .35s
+    &.active
+      visibility: visible
+      opacity: 1
 
 .chapter-nav__wrap
-  position: absolute
-  left: 1rem
-  top: 0
+  @media (min-width: 576px)
+    position: absolute
+    left: 2rem
+    top: 0
 
 .chapter-link
   background: none
@@ -34,3 +54,17 @@
   list-style-type: none
   padding-left: 0
   margin-bottom: 0
+
+.chapter-nav__item:not(:last-child)
+  margin-bottom: 1rem
+  @media (min-width: 576px)
+    margin-bottom: 0
+
+.chapter-nav__list--visible
+  margin-top: 1rem
+  @media (min-width: 576px)
+    margin-top: 0
+  .chapter-nav__item
+    display: inline-block
+    @media (min-width: 576px)
+      display: block
