+++ sass/layout/_general.sass
@@ -1,5 +1,7 @@
 body
   color: $grey-mid
+  &.no-scroll
+    overflow-y: hidden
 
 p
   font-family: $stack-font
@@ -11,3 +13,7 @@ h4,
 h5
   color: $black
   margin-bottom: 2rem
+
+img
+  max-width: 100%
+  height: auto
