+++ sass/layout/_sidebar.sass
@@ -1,2 +1,3 @@
 .col--nav
-  flex: 0 0 370px
+  @media (min-width: 992px)
+    flex: 0 0 370px
